+++ docs/CHANGELOG.md
@@ -3,7 +3,9 @@
 ### v0.3.3 (dev)
 - Extended `Equal` function to all expression types
 - Added functions:
+  - Matrix: `Reshape`
   - Set: `Difference`, `Powerset`, `Subset`, `SymmetricDifference`
+- Modified existing functions to support new expression types
 ### v0.3.2 (Jan 17, 2022)
 - Set up generic compiler to generalize classes
 - Simplified step evaluation process